Refactor passing table ref & Change pagination styles
Before, table ref was created within useStickyTableHead and returned
from there. It was retrieved within CoinsTable component
to pass it to table element for reference where sticky table head
should appear.
Then, to make scrolling to top of the table possible when changing
page, table ref was needed in Coins component so it can then be
passed with handlePagination function to Pagination component.
It was being retrieved from CoinsTable using useState and handler
function.
All of this was a little bit messy.

Now table ref is created within Coins component and passed further
to components and hooks that need it.

Additionally:
- Pagination was moved to a seperate component

- Media queries were added to make pagination and entire page
better responding to small (mobile) screens